requirements.txt
discord.py==1.3.3
youtube_dl==2020.6.16.1
asyncio==3.4.3
python-dotenv==0.13.0
PyNaCl==1.4.0